requirements.txt
faiss-cpu
javalang
numpy
openai
pandas
pygments
torch
tqdm
transformers
chardet
regex
